all: updated new features
Added kubernetes resource usage and display to user if the the pod is
highlighted.

Can now go up and down with the arrow keys to change what is
highlighted.

SPACE will take a snapshot of the current data and compare that to all
new, updated data as it comes in from the metrics client.